PDFBOX-4892: LGTM fix
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1897264 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/interactive/annotation/handlers/CloudyBorder.java
@@ -85,7 +85,7 @@ class CloudyBorder
      */
     void createCloudyRectangle(PDRectangle rd) throws IOException
     {
-        rectWithDiff = applyRectDiff(rd, lineWidth / 2);
+        rectWithDiff = applyRectDiff(rd, (float) (lineWidth / 2));
         double left = rectWithDiff.getLowerLeftX();
         double bottom = rectWithDiff.getLowerLeftY();
         double right = rectWithDiff.getUpperRightX();
@@ -493,7 +493,7 @@ class CloudyBorder
         }
     }
 
-    private PDRectangle applyRectDiff(PDRectangle rd, double min)
+    private PDRectangle applyRectDiff(PDRectangle rd, float min)
     {
         float rectLeft = annotRect.getLowerLeftX();
         float rectBottom = annotRect.getLowerLeftY();
@@ -506,10 +506,10 @@ class CloudyBorder
         rectRight = Math.max(rectLeft, rectRight);
         rectTop = Math.max(rectBottom, rectTop);
 
-        double rdLeft;
-        double rdBottom;
-        double rdRight;
-        double rdTop;
+        float rdLeft;
+        float rdBottom;
+        float rdRight;
+        float rdTop;
 
         if (rd != null)
         {
